move some defaults calls to defaults service
