+++ OnrampProject/ViewModels/WeatherViewModel.swift
@@ -21,7 +21,7 @@ class WeatherViewModel {
     // MARK: - Properties
     
     private let weatherNetworkService = WeatherNetworkService()
-    fileprivate let userDefaults = UserDefaults.standard
+    fileprivate let userDefaultsService = UserDefaultsService()
     private(set) var weatherCity: String
     private(set) var weatherState: String
     private(set) var lastUpdateTime: String?
@@ -36,7 +36,7 @@ class WeatherViewModel {
             
             lastUpdateTime = currentTimeAsString
             
-            WeatherViewModel.storeInUserDefaults(item: weather)
+            userDefaultsService.storeInUserDefaults(item: weather)
             delegate?.didUpdateWeather(self)
         }
     }
@@ -79,7 +79,7 @@ class WeatherViewModel {
         self.weatherState = state
         modelImageViewModel.delegate = self
         
-        if let lastKnownWeather = WeatherViewModel.getFromUserDefaults(item: Constants.userDefaultKeys.lastKnownWeather) as? Weather {
+        if let lastKnownWeather = userDefaultsService.getFromUserDefaults(item: Constants.userDefaultKeys.lastKnownWeather) as? Weather {
             self.weather = lastKnownWeather
             lastUpdateTime = self.weather?.date.getTimeAsString()
             delegate?.didUpdateWeather(self)
@@ -102,7 +102,7 @@ class WeatherViewModel {
                 
                 self.weather = Weather(date: date, temperature: Int(data.main.temp), hourlyTemps: [:], city: city, state: state, currentHourBlock: currentHourBlock)
                 
-                WeatherViewModel.storeInUserDefaults(item: self.weather)
+                self.userDefaultsService.storeInUserDefaults(item: self.weather)
             }
         }
     }
@@ -156,69 +156,5 @@ extension Date {
 extension WeatherViewModel: ModelImageViewModelDelegate {
     func modelName(_ modelImageViewModel: ModelImageViewModel, didChange: Bool) {
         delegate?.didUpdateModelImageDetails(self, modelImageViewModel: modelImageViewModel)
-    }
-}
-
-// TODO: Move to userdefault service or shared instance
-extension WeatherViewModel {
-    
-    static func storeInUserDefaults(item: Any?) {
-        let userDefaults = UserDefaults.standard
-        
-        if let item = item as? Weather {
-            let encoder = JSONEncoder()
-            
-            do {
-                let data = try encoder.encode(item)
-                userDefaults.set(data, forKey: Constants.userDefaultKeys.lastKnownWeather)
-            } catch {
-                print("error storing data in user defaults: \(error)")
-            }
-        }
-        
-        if let item = item as? Settings {
-            let encoder = JSONEncoder()
-            
-            do {
-                let data = try encoder.encode(item)
-                userDefaults.set(data, forKey: Constants.userDefaultKeys.settings)
-            } catch {
-                print("error storing data in user defaults: \(error)")
-            }
-        }
-    }
-    
-    static func getFromUserDefaults(item: String) -> Any? {
-        let userDefaults = UserDefaults.standard
-        let decoder = JSONDecoder()
-        
-        guard let data = userDefaults.data(forKey: item) else {
-            print("items has not been stored in defaults yet")
-            return nil
-        }
-        
-        if item == Constants.userDefaultKeys.lastKnownWeather {
-            do {
-                
-                let lastKnownWeather = try decoder.decode(Weather.self, from: data)
-                return lastKnownWeather
-            } catch {
-                print("error retrieving last known weather in user defaults: \(error)")
-                return nil
-            }
-        }
-        
-        if item == Constants.userDefaultKeys.settings {
-            do {
-                let data = userDefaults.object(forKey: item) as! Data
-                let settings = try decoder.decode(Settings.self, from: data)
-                return settings
-            } catch {
-                print("error retrieving settings in user defaults")
-                return nil
-            }
-        }
-        
-        return nil
     }
 }
